third_party/blink/web_tests/external/wpt/editing/run/insertunorderedlist_1001-2000-expected.txt

This is a testharness.js-based test.
Found 200 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<ul><li>foo<li>bar</li></li></ul>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><li>bar</li></ul><div>baz</div><p>extra</p>" but got "foo<br>bar<br>baz<br><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><li>bar</li></ul><p>baz</p><p>extra</p>" but got "foo<br>bar<br>baz<br><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<ul><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "foo<ul><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li><li>baz</li></ul>" but got "<ul><li>foo<li>bar</li>baz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><div>baz</div>" but got "foo<ul><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><p>baz</p>" but got "foo<ul><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul><li>bar</li><li>baz</li></ul>" but got "foo<br>bar<br>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul><li>bar</li><li>baz</li></ul>" but got "foo<br>bar<br>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandState("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandIndeterm("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li></ol><div>baz</div>quz" but got "foo<ol><li>bar</li></ol>baz<br>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li></ol><p>baz</p>quz" but got "foo<ol><li>bar</li></ol>baz<br>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<ol><li>bar</li></ol><ul><li>baz</li><li>quz</li></ul>" but got "foo<ol><li>bar</li></ol>baz<br><ul><li>quz</li></ul>"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div>bar</div><ol><li>baz</li></ol>quz" but got "foo<br>bar<br><ol><li>baz</li></ol>quz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>bar</p><ol><li>baz</li></ol>quz" but got "foo<br>bar<br><ol><li>baz</li></ol>quz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul><ol><li>baz</ol>quz" queryCommandIndeterm("insertunorderedlist") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<ul><li>bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo]<blockquote>bar</blockquote>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo]<blockquote>bar</blockquote>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo<blockquote>[bar]</blockquote>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<blockquote>[bar]</blockquote>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul>baz" but got "<ul><li>foo</li><li>bar</li></ul>baz"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]</blockquote>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]</blockquote>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo]<blockquote><ol><li>bar</ol></blockquote>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo]<blockquote><ol><li>bar</ol></blockquote>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "foo<blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "foo<blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><blockquote>baz</blockquote>" but got "<ul><li>foo</li><li>bar</li></ul><blockquote>baz</blockquote>"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><blockquote><p>bar</p></blockquote><p>baz</p>" but got "<p><ul><li>foo</li></ul></p><blockquote><p>bar</p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><blockquote><p>bar</p></blockquote><p>baz</p>" but got "<p><ul><li>foo</li></ul></p><blockquote><p>bar</p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><blockquote><ul><li>bar</li></ul></blockquote><p>baz</p>" but got "<p>foo</p><blockquote><p><ul><li>bar</li></ul></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><blockquote><ul><li>bar</li></ul></blockquote><p>baz</p>" but got "<p>foo</p><blockquote><p><ul><li>bar</li></ul></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><p>baz</p>" but got "<p><ul><li>foo</li><li>bar</li></ul></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote><ul><li>bar</li></ul></blockquote><p>baz</p>" but got "<ol><li>foo</li></ol><blockquote><p><ul><li>bar</li></ul></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote><ul><li>bar</li></ul></blockquote><p>baz</p>" but got "<ol><li>foo</li></ol><blockquote><p><ul><li>bar</li></ul></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul id=\\"abc\\"><li>foo</li></ul><div>bar</div><ul><li>baz</li></ul>" but got "<ul id=\\"abc\\"><li>foo</li></ul>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul id=\\"abc\\"><li>foo</li></ul><p>bar</p><ul><li>baz</li></ul>" but got "<ul id=\\"abc\\"><li>foo</li></ul>bar<br><ul><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><div>bar</div><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><span style=\\"color:rgb(0, 0, 255)\\">bar</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><div>bar</div><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><span style=\\"color:rgb(0, 0, 255)\\">bar</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><p>bar</p><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><span style=\\"color:rgb(0, 0, 255)\\">bar</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><p>bar</p><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><span style=\\"color:rgb(0, 0, 255)\\">bar</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><div>bar</div><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><span style=\\"text-indent:1em\\">bar</span><br><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><div>bar</div><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><span style=\\"text-indent:1em\\">bar</span><br><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><p>bar</p><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><span style=\\"text-indent:1em\\">bar</span><br><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><p>bar</p><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><span style=\\"text-indent:1em\\">bar</span><br><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul id=\\"abc\\"><li>bar</li><li>baz</li></ul>" but got "foo<br><ul id=\\"abc\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul id=\\"abc\\"><li>bar</li><li>baz</li></ul>" but got "foo<br><ul id=\\"abc\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foo</span><br><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>" but got "<span style=\\"text-indent:1em\\">foo</span><br><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>" but got "<span style=\\"text-indent:1em\\">foo</span><br><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul id=\\"abc\\"><li>foo</li><li>bar</li></ul><div>baz</div>" but got "<ul id=\\"abc\\"><li>foo</li><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul id=\\"abc\\"><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<ul id=\\"abc\\"><li>foo</li><li>bar</li></ul>baz"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><div>baz</div>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><span style=\\"color:rgb(0, 0, 255)\\">baz</span>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><span style=\\"color:rgb(0, 0, 255)\\">baz</span>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><div>baz</div>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><span style=\\"text-indent:1em\\">baz</span>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><p>baz</p>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><span style=\\"text-indent:1em\\">baz</span>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li></ul>" but got "<ul><li>foo</li><li>bar</li></ul> "
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li></ul>" but got "<ul><li>foo</li><li>bar</li></ul> "
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li></ul>" but got " <ul><li>foo</li><li>bar</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li></ul>" but got " <ul><li>foo</li><li>bar</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<p>[foo]</p> <ul><li>bar</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li> <li>baz</li></ul>" but got "<ul><li>foo</li><li>bar</li><li>baz</li></ul>  "
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li> <li>bar</li> <li>baz</li></ul>" but got "<ul><li>foo</li><li>bar</li><li>baz</li></ul>  "
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "<ul><li>foo</ul> <p>[bar]</p> <ul><li>baz</ul>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li style=\\"font-size:1.3em\\">1</li><li style=\\"font-size:1.1em\\">2</li></ul>" but got "<div style=\\"font-size:1.3em\\"><ul><li>1</li><li>2</li></ul></div>"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li style=\\"font-size:1.3em\\">1</li><li style=\\"font-size:1.1em\\">2</li></ul>" but got "<div style=\\"font-size:1.3em\\"><ul><li>1</li><li>2</li></ul></div>"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertunorderedlist",""]] "{<div style=\\"font-size: 1.3em\\">1</div><div style=\\"font-size: 1.1em\\">2</div>}" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}</div>": execCommand("insertunorderedlist", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}</div>": execCommand("inserttext", false, "abc") return value
  assert_equals: expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div></div>" not in array ["<div><ul><li>abc</li></ul></div>", "<div><ul><li>abc<br></li></ul></div>"]
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}</div>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}</div>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "true" but got "false"
Harness: the test ran to completion.
